backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlmodel>=0.0.22
anthropic>=0.40.0
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.1
apscheduler>=3.10.4
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
# Production additions
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
httpx>=0.27.0
